requirements.txt
django
djangorestframework==3.12.*
